# internal/codegen/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /and-orf7kpd)
	ComplexAuthenticationG9wTc(w http.ResponseWriter, r *http.Request)

	// (GET /andt2mvx)
	CombinedAuthenticationA6rDh(w http.ResponseWriter, r *http.Request)

	// (GET /api-key-cookieh8ryu)
	ApiKeyCookieD4tZu(w http.ResponseWriter, r *http.Request)

	// (GET /api-key-headerx3nws)
	ApiKeyHeaderL5mQa(w http.ResponseWriter, r *http.Request)

	// (GET /api-key-queryc6jfa)
	ApiKeyQueryW7pEr(w http.ResponseWriter, r *http.Request)

	// (GET /api-key-with-rolek2vdp)
	ApiKeyWithRoleP9bLc(w http.ResponseWriter, r *http.Request)

	// (GET /basicn5sge)
	BasicAuthenticationE6qRt(w http.ResponseWriter, r *http.Request)

	// (GET /bearerw7cqm)
	BearerAuthenticationT3nVk(w http.ResponseWriter, r *http.Request)

	// (GET /custom-httpb4xlt)
	CustomHttpAuthenticationH8mYs(w http.ResponseWriter, r *http.Request)

	// (GET /inherited-securitya8k2z)
	InheritedSecurityF3vNx(w http.ResponseWriter, r *http.Request)

	// (GET /mutual-tlsj9rke)
	MutualTlsAuthenticationK4pBx(w http.ResponseWriter, r *http.Request)

	// (GET /oauth2-without-scopesd3wnf)
	Oauth2WithoutScopesB7vQe(w http.ResponseWriter, r *http.Request)

	// (GET /oauth2z6pua)
	Oauth2AuthenticationM2cFw(w http.ResponseWriter, r *http.Request)

	// (GET /openid-connecty8hsb)
	OpenIdConnectAuthenticationS5jLu(w http.ResponseWriter, r *http.Request)

	// (GET /optional-authenticationp9tql)
	OptionalAuthenticationC2yHs(w http.ResponseWriter, r *http.Request)

	// (GET /orq5lzc)
	AlternativeAuthenticationV8xNp(w http.ResponseWriter, r *http.Request)

	// (GET /publicm4vrb)
	PublicEndpointR8kWd(w http.ResponseWriter, r *http.Request)

	// (GET /referenced-security-schemer4cnj)
	ReferencedSecuritySchemeU2kMf(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// (GET /and-orf7kpd)
func (_ Unimplemented) ComplexAuthenticationG9wTc(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /andt2mvx)
func (_ Unimplemented) CombinedAuthenticationA6rDh(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api-key-cookieh8ryu)
func (_ Unimplemented) ApiKeyCookieD4tZu(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api-key-headerx3nws)
func (_ Unimplemented) ApiKeyHeaderL5mQa(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api-key-queryc6jfa)
func (_ Unimplemented) ApiKeyQueryW7pEr(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api-key-with-rolek2vdp)
func (_ Unimplemented) ApiKeyWithRoleP9bLc(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /basicn5sge)
func (_ Unimplemented) BasicAuthenticationE6qRt(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /bearerw7cqm)
func (_ Unimplemented) BearerAuthenticationT3nVk(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /custom-httpb4xlt)
func (_ Unimplemented) CustomHttpAuthenticationH8mYs(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /inherited-securitya8k2z)
func (_ Unimplemented) InheritedSecurityF3vNx(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /mutual-tlsj9rke)
func (_ Unimplemented) MutualTlsAuthenticationK4pBx(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /oauth2-without-scopesd3wnf)
func (_ Unimplemented) Oauth2WithoutScopesB7vQe(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /oauth2z6pua)
func (_ Unimplemented) Oauth2AuthenticationM2cFw(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /openid-connecty8hsb)
func (_ Unimplemented) OpenIdConnectAuthenticationS5jLu(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /optional-authenticationp9tql)
func (_ Unimplemented) OptionalAuthenticationC2yHs(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /orq5lzc)
func (_ Unimplemented) AlternativeAuthenticationV8xNp(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /publicm4vrb)
func (_ Unimplemented) PublicEndpointR8kWd(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /referenced-security-schemer4cnj)
func (_ Unimplemented) ReferencedSecuritySchemeU2kMf(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ComplexAuthenticationG9wTc operation middleware
func (siw *ServerInterfaceWrapper) ComplexAuthenticationG9wTc(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ComplexAuthenticationG9wTc(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CombinedAuthenticationA6rDh operation middleware
func (siw *ServerInterfaceWrapper) CombinedAuthenticationA6rDh(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CombinedAuthenticationA6rDh(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ApiKeyCookieD4tZu operation middleware
func (siw *ServerInterfaceWrapper) ApiKeyCookieD4tZu(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ApiKeyCookieD4tZu(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ApiKeyHeaderL5mQa operation middleware
func (siw *ServerInterfaceWrapper) ApiKeyHeaderL5mQa(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ApiKeyHeaderL5mQa(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ApiKeyQueryW7pEr operation middleware
func (siw *ServerInterfaceWrapper) ApiKeyQueryW7pEr(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ApiKeyQueryW7pEr(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ApiKeyWithRoleP9bLc operation middleware
func (siw *ServerInterfaceWrapper) ApiKeyWithRoleP9bLc(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ApiKeyWithRoleP9bLc(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// BasicAuthenticationE6qRt operation middleware
func (siw *ServerInterfaceWrapper) BasicAuthenticationE6qRt(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.BasicAuthenticationE6qRt(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// BearerAuthenticationT3nVk operation middleware
func (siw *ServerInterfaceWrapper) BearerAuthenticationT3nVk(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.BearerAuthenticationT3nVk(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CustomHttpAuthenticationH8mYs operation middleware
func (siw *ServerInterfaceWrapper) CustomHttpAuthenticationH8mYs(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CustomHttpAuthenticationH8mYs(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// InheritedSecurityF3vNx operation middleware
func (siw *ServerInterfaceWrapper) InheritedSecurityF3vNx(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.InheritedSecurityF3vNx(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// MutualTlsAuthenticationK4pBx operation middleware
func (siw *ServerInterfaceWrapper) MutualTlsAuthenticationK4pBx(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.MutualTlsAuthenticationK4pBx(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Oauth2WithoutScopesB7vQe operation middleware
func (siw *ServerInterfaceWrapper) Oauth2WithoutScopesB7vQe(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Oauth2WithoutScopesB7vQe(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Oauth2AuthenticationM2cFw operation middleware
func (siw *ServerInterfaceWrapper) Oauth2AuthenticationM2cFw(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Oauth2AuthenticationM2cFw(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// OpenIdConnectAuthenticationS5jLu operation middleware
func (siw *ServerInterfaceWrapper) OpenIdConnectAuthenticationS5jLu(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.OpenIdConnectAuthenticationS5jLu(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// OptionalAuthenticationC2yHs operation middleware
func (siw *ServerInterfaceWrapper) OptionalAuthenticationC2yHs(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.OptionalAuthenticationC2yHs(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// AlternativeAuthenticationV8xNp operation middleware
func (siw *ServerInterfaceWrapper) AlternativeAuthenticationV8xNp(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AlternativeAuthenticationV8xNp(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PublicEndpointR8kWd operation middleware
func (siw *ServerInterfaceWrapper) PublicEndpointR8kWd(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PublicEndpointR8kWd(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ReferencedSecuritySchemeU2kMf operation middleware
func (siw *ServerInterfaceWrapper) ReferencedSecuritySchemeU2kMf(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ReferencedSecuritySchemeU2kMf(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/inherited-securitya8k2z", wrapper.InheritedSecurityF3vNx)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/publicm4vrb", wrapper.PublicEndpointR8kWd)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/optional-authenticationp9tql", wrapper.OptionalAuthenticationC2yHs)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api-key-headerx3nws", wrapper.ApiKeyHeaderL5mQa)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api-key-queryc6jfa", wrapper.ApiKeyQueryW7pEr)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api-key-cookieh8ryu", wrapper.ApiKeyCookieD4tZu)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api-key-with-rolek2vdp", wrapper.ApiKeyWithRoleP9bLc)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/basicn5sge", wrapper.BasicAuthenticationE6qRt)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/bearerw7cqm", wrapper.BearerAuthenticationT3nVk)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/custom-httpb4xlt", wrapper.CustomHttpAuthenticationH8mYs)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/mutual-tlsj9rke", wrapper.MutualTlsAuthenticationK4pBx)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/oauth2z6pua", wrapper.Oauth2AuthenticationM2cFw)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/oauth2-without-scopesd3wnf", wrapper.Oauth2WithoutScopesB7vQe)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/openid-connecty8hsb", wrapper.OpenIdConnectAuthenticationS5jLu)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/orq5lzc", wrapper.AlternativeAuthenticationV8xNp)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/andt2mvx", wrapper.CombinedAuthenticationA6rDh)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/and-orf7kpd", wrapper.ComplexAuthenticationG9wTc)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/referenced-security-schemer4cnj", wrapper.ReferencedSecuritySchemeU2kMf)
	})

	return r
}